lavc/aac_ac3_parser: fix the potential overflow
Fix the potential overflow.

Suggested-by: Alexander Strasser <eclipse7@gmx.net>

+++ libavcodec/aac_ac3_parser.h
@@ -55,8 +55,8 @@ typedef struct AACAC3ParseContext {
     uint64_t state;
 
     int need_next_header;
-    int frame_number;
-    int last_bit_rate;
+    uint64_t frame_number;
+    double last_bit_rate;
     enum AVCodecID codec_id;
 } AACAC3ParseContext;
